dsrewr

+++ 0x15-file_io/1-create_file.c
@@ -10,7 +10,7 @@
 */
 int create_file(const char *filename, char *text_content)
 {
-    ssize_t fileToCreate;
+    int fileToCreate;
     ssize_t writeText;
 
     if (filename == NULL)
